requirements.txt
praw ==7.4.0
newsapi-python == 0.2.7
matplotlib
python-dotenv